requirements.txt
pandas==1.4.0
Pillow 
scikit-learn==1.0.2
streamlit 
urllib3 
watchdog 
